chore(suite): reorganize transaction components
