+++ packages/suite/src/components/wallet/TransactionItem/CommonComponents.tsx
@@ -1,7 +1,7 @@
 import styled, { css } from 'styled-components';
 import { variables } from '@trezor/components';
 import { FormattedCryptoAmount, FormattedNftAmount, HiddenPlaceholder } from 'src/components/suite';
-import { MIN_ROW_HEIGHT } from './BaseTargetLayout';
+import { MIN_ROW_HEIGHT } from './TransactionTargetLayout';
 
 export const TxTypeIconWrapper = styled.div`
     padding-right: 24px;
