+++ packages/suite/src/components/wallet/TransactionItem/TransactionTypeIcon.tsx
