+++ packages/suite/src/views/wallet/transactions/TransactionList/TransactionCandidates.tsx
@@ -4,8 +4,8 @@ import { Card, SVG_IMAGES, variables } from '@trezor/components';
 import { useSelector } from 'src/hooks/suite';
 import { selectCoinjoinAccountByKey } from 'src/reducers/wallet/coinjoinReducer';
 import { SECONDARY_PANEL_HEIGHT } from 'src/components/suite/AppNavigation';
-import { TransactionTypeIcon } from 'src/components/wallet/TransactionItem/components/TransactionTypeIcon';
-import { TxTypeIconWrapper } from 'src/components/wallet/TransactionItem/components/CommonComponents';
+import { TransactionTypeIcon } from 'src/components/wallet/TransactionItem/TransactionTypeIcon';
+import { TxTypeIconWrapper } from 'src/components/wallet/TransactionItem/CommonComponents';
 import { Translation } from 'src/components/suite/Translation';
 import TooltipSymbol from 'src/components/suite/TooltipSymbol';
