+++ packages/suite/src/views/wallet/transactions/TransactionList/TransactionListActions/TransactionListActions.tsx
@@ -1,4 +1,5 @@
 import styled from 'styled-components';
+
 import { SearchAction, SearchProps } from './SearchAction';
 import { ExportAction, ExportActionProps } from './ExportAction';
